+++ web/src/components/CostAllocation.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { Editor } from 'react-draft-wysiwyg';
 
-import Icon, { faHelp } from '../Icons';
-import { EDITOR_CONFIG } from '../../util/editor';
+import Icon, { faHelp } from './Icons';
+import { EDITOR_CONFIG } from '../util';
 
 class CostAllocation extends Component {
   state = {
