+++ web/src/components/HelpBox.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import Icon, { faHelpSolid } from '../Icons';
+import Icon, { faHelpSolid } from './Icons';
 
 const HelpBox = ({ children }) => (
   <div className="my2 p2 h5 sm-col-7 lg-col-6 bg-teal-light border border-teal border-width-2 rounded relative">
