--- web/src/components/SubmissionSuccess.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const SubmissionSuccess = () => (
-  <div>
-    <h1>Thanks! We’ve received your request.</h1>
-    <p>Some sort of confirmation number or note about checking email etc.</p>
-    <p>A sentence about how long it usually takes to receive a response.</p>
-    <p>A sentence about what to do if you have questions.</p>
-  </div>
-);
-
-export default SubmissionSuccess;
