+++ web/src/util/index.js
@@ -111,6 +111,32 @@ export const STANDARDS = [
   }
 ];
 
+export const EDITOR_CONFIG = {
+  options: ['inline', 'blockType', 'fontSize', 'image', 'list', 'link'],
+  inline: {
+    options: ['bold', 'italic', 'underline']
+  },
+  image: {
+    className: undefined,
+    component: undefined,
+    popupClassName: undefined,
+    urlEnabled: true,
+    uploadEnabled: true,
+    alignmentEnabled: true,
+    uploadCallback: undefined,
+    previewImage: false,
+    inputAccept: 'image/gif,image/jpeg,image/jpg,image/png,image/svg',
+    alt: { present: false, mandatory: false },
+    defaultSize: {
+      height: 'auto',
+      width: 'auto'
+    }
+  },
+  list: {
+    options: ['unordered', 'ordered']
+  }
+};
+
 export const stateLookup = id => STATES.find(s => s.id === id.toLowerCase());
 
 export const getParams = str =>
